+++ bypy/bypy.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python -tt
 # encoding: utf-8
 # PYTHON_ARGCOMPLETE_OK
 # ===  IMPORTANT  ====
@@ -113,6 +113,8 @@ except:
 		perr("Something seems wrong with the urllib3 installation.\nQuitting")
 		sys.exit(const.EFatal)
 
+from requests_toolbelt import multipart
+
 # http://stackoverflow.com/a/27320254/404271
 try:
 	import multiprocess as mp
@@ -135,6 +137,10 @@ if Pool != None:
 
 # global instance for non-member function to access
 gbypyinst = None
+
+pcsurl = const.PcsUrl
+cpcsurl = const.CPcsUrl
+dpcsurl = const.DPcsUrl
 
 # This crap is here to avoid circular imports
 # What a fantastic packing architecture Python has!
@@ -295,6 +301,7 @@ class ByPy(object):
 		mirror = '',
 		selectmirror = False,
 		resumedl_revertcount = const.DefaultResumeDlRevertCount,
+		deletesource = False,
 		verbose = 0, debug = False,
 		configdir = const.ConfigDir,
 		requester = RequestsRequester,
@@ -364,6 +371,10 @@ class ByPy(object):
 		self.__followlink = followlink
 		self.__rapiduploadonly = rapiduploadonly
 		self.__resumedl_revertcount = resumedl_revertcount
+		self.__deletesource = deletesource
+		if deletesource:
+			self.pd("Forcing verification since we will delete source for successful transfers.")
+			self.__verify = True
 		self.processes = processes
 
 		# these two variables are without leadning double underscaore "__" as to export the as public,
@@ -380,17 +391,9 @@ class ByPy(object):
 		mirror = mirror.lower()
 		if len(mirror) > 0 and mirror != const.PcsDomain:
 			self.setpcsurl(mirror)
-		else:
-			if selectmirror:
-				self.__select_fastest_mirror()
-			else:
-				# use the default domain
-				global pcsurl
-				global cpcsurl
-				global dpcsurl
-				pcsurl = const.PcsUrl
-				cpcsurl = const.CPcsUrl
-				dpcsurl = const.DPcsUrl
+		elif selectmirror:
+			self.__select_fastest_mirror()
+		# else: already set in global
 
 		if self.__checkssl:
 			# sort of undocumented by requests
@@ -539,19 +542,41 @@ class ByPy(object):
 		errors = list(filter(lambda x: x != const.ENoError, results))
 		return const.ENoError if len(errors) == 0 else errors[-1]
 
-	def __multi_process(self, worker, iter, process = "dl / ul"):
+	def __multi_process(self, worker, iterator, process = "dl / ul"):
 		if not self.__check_prompt_multiprocess():
 			return const.EArgument
 		self.__warn_multi_processes(process)
 		# patch the printer
 		restoremp = set_mp_print()
 		with UPool(self.processes) as pool:
-			# http://stackoverflow.com/a/35134329/404271
-			# On Python 2.x, Ctrl-C won't work if we use pool.map(), wait() or get()
-			ar = pool.map_async(worker, iter)
-			results = ar.get(const.TenYearInSeconds)
-			restoremp()
-			return self.__filter_multi_results(results)
+			try:
+				# http://stackoverflow.com/a/35134329/404271
+				# On Python 2.x, Ctrl-C won't work if we use pool.map(), wait() or get()
+				ar = pool.map_async(worker, iterator)
+				#results = ar.get(const.TenYearInSeconds)
+				results = ar.get(const.FortyNineDaysInSeconds)
+				restoremp()
+				return self.__filter_multi_results(results)
+			except pickle.PicklingError as pe:
+				errmsg = ("Exception:\n{}\n"
+						"--------------------------------\n"
+						"Error: Your Python 'multiprocess' library is probably "
+						"not properly installed (missing C extensions). "
+						"You need to install a C compiler and Python headers before "
+						"installing the Python 'multiprocess' library. "
+						"(All these hassles could have been saved if Python's builtin "
+						"'multiprocessing' works properly, sigh.)\n"
+						"Fix for debian derivatives:\n"
+						"- Install gcc: # apt-get install gcc\n"
+						"- Install python-dev: # apt-get install python-dev\n"
+						"- Reinstall Python 'multiprocess' library:\n"
+						"  # pip uninstall -y multiprocess\n"
+						"  # pip install -v multiprocess\n"
+						"- If there's no errors/warnings in the above actions, "
+						"then this error should be gone when you run 'bypy' with '{}' again.\n"
+						).format(formatex(pe), const.MultiprocessOption)
+				perr(errmsg)
+				self.quit(const.EFatal)
 
 	def __print_error_json(self, r):
 		try:
@@ -1091,6 +1116,26 @@ Possible fixes:
 		''' Usage: refreshtoken - refresh the access token '''
 		return self.__refresh_token()
 
+	def __remove_remote_on_success(self, remotepath):
+		if self.__deletesource:
+			self.pd("Removing remote path '{}' after successful download.".format(remotepath))
+			result = self.__delete(remotepath)
+			if result == const.ENoError:
+				self.pd("Remote path '{}' removed.".format(remotepath))
+			else:
+				perr("Failed to remove remote path '{}'.".format(remotepath))
+			return result
+
+	def __remove_local_on_success(self, localpath):
+		if self.__deletesource:
+			self.pd("Removing local path '{}' after successful upload.".format(localpath))
+			result = cachedm.remove_path_and_cache(localpath)
+			if result == const.ENoError:
+				self.pd("Local path '{}' removed.".format(localpath))
+			else:
+				perr("Failed to remove local path '{}'.".format(localpath))
+			return result
+
 	def info(self):
 		return self.quota()
 
@@ -1346,16 +1391,38 @@ get information of the given path (dir / file) at Baidu Yun.
 			perr("'{}' >>==> '{}' FAILED.".format(self.__current_file, args))
 			return const.EHashMismatch
 
+	# want to be proper? properness doesn't work
+	# there seems to be a bug at Baidu's handling of http text:
+	# Content-Disposition: ...  filename=utf-8''yourfile.ext
+	# (pass '-ddd' to this program to verify this)
+	# when you specify a unicode file name, which will be encoded
+	# using the utf-8'' syntax
+	# so, we put a work-around here: we always call our file 'file'
+	# NOTE: an empty file name '' doesn't seem to work, so we
+	# need to give it a name at will, but empty one.
+	# apperantly, Baidu PCS doesn't use this file name for
+	# checking / verification, so we are probably safe here.
+	#files = { 'file' : (os.path.basename(self.__current_file), self.__current_slice) } )
+	#files = { 'file' : (os.path.basename(localpath), f) })
+	#files = { 'file' : ('file', f) })
+	def __stream_upload(self, fo, pars, act, remotepath,
+			url = cpcsurl + 'file',
+			uploadfilename = 'file'):
+		# https://stackoverflow.com/a/35784072/404271
+		form = multipart.encoder.MultipartEncoder({
+			'file': (uploadfilename, fo, 'application/octet-stream')})
+		headers = {"Content-Type": form.content_type}
+		return self.__post(url,
+			pars, act, remotepath,
+			headers=headers, data=form)
+
 	def __upload_slice(self, remotepath):
 		pars = {
 			'method' : 'upload',
 			'type' : 'tmpfile'}
 
-		return self.__post(cpcsurl + 'file',
-				pars, self.__upload_slice_act, remotepath,
-				# want to be proper? properness doesn't work (search this sentence for more occurence)
-				#files = { 'file' : (os.path.basename(self.__current_file), self.__current_slice) } )
-				files = { 'file' : ('file', self.__current_slice) } )
+		return self.__stream_upload(self.__current_slice,
+				pars, self.__upload_slice_act, remotepath)
 
 	def __update_progress_entry(self, fullpath):
 		progress = jsonload(const.ProgressPath)
@@ -1524,21 +1591,8 @@ get information of the given path (dir / file) at Baidu Yun.
 		self.__add_isrev_param(ondup, pars)
 
 		with io.open(localpath, 'rb') as f:
-			return self.__post(cpcsurl + 'file',
-				pars, self.__upload_one_file_act, remotepath,
-				# wants to be proper? properness doesn't work
-				# there seems to be a bug at Baidu's handling of http text:
-				# Content-Disposition: ...  filename=utf-8''yourfile.ext
-				# (pass '-ddd' to this program to verify this)
-				# when you specify a unicode file name, which will be encoded
-				# using the utf-8'' syntax
-				# so, we put a work-around here: we always call our file 'file'
-				# NOTE: an empty file name '' doesn't seem to work, so we
-				# need to give it a name at will, but empty one.
-				# apperantly, Baidu PCS doesn't use this file name for
-				# checking / verification, so we are probably safe here.
-				#files = { 'file' : (os.path.basename(localpath), f) })
-				files = { 'file' : ('file', f) })
+			return self.__stream_upload(f, pars,
+					self.__upload_one_file_act, remotepath)
 
 	#TODO: upload empty directories as well?
 	def __walk_upload(self, localpath, remotepath, ondup, walk):
@@ -1606,16 +1660,21 @@ get information of the given path (dir / file) at Baidu Yun.
 
 	def __upload_dir(self, localpath, remotepath, ondup = 'overwrite'):
 		self.pd("Uploading directory '{}' to '{}'".format(localpath, remotepath))
+		result = const.ENoError
 		if Pool and self.processes > 1:
-			return self.__upload_dir_multi(localpath, remotepath, ondup)
+			result = self.__upload_dir_multi(localpath, remotepath, ondup)
 		else:
-			return self.__upload_dir_single(localpath, remotepath, ondup)
+			result =  self.__upload_dir_single(localpath, remotepath, ondup)
+		if result == const.ENoError:
+			self.__remove_local_on_success(localpath)
+		return result
 
 	def __upload_file(self, localpath, remotepath, ondup = 'overwrite'):
 		# TODO: this is a quick patch
 		if not self.__shallinclude(localpath, remotepath, True):
 			# since we are not going to upload it, there is no error
-			return const.ENoError
+			#return const.ENoError
+			return const.ESkipped
 
 		self.__current_file = localpath
 		self.__current_file_size = getfilesize(localpath)
@@ -1648,12 +1707,18 @@ get information of the given path (dir / file) at Baidu Yun.
 							self.__current_file_size))
 				else:
 					self.pv("'{}' can't be rapidly uploaded, so it's skipped since we are in the rapid-upload-only mode.".format(localpath))
-
-			return result
-		elif not self.__rapiduploadonly:
+					result = const.ESkipped
+		elif self.__rapiduploadonly:
+			self.pv("'{}' is too small to be rapidly uploaded, so it's skipped since we are in the rapid-upload-only mode.".format(localpath))
+			result = const.ESkipped
+		else:
 			# very small file, must be uploaded manually and no slicing is needed
 			self.pd("'{}' is small and being non-slicing uploaded.".format(self.__current_file))
-			return self.__upload_one_file(localpath, remotepath, ondup)
+			result = self.__upload_one_file(localpath, remotepath, ondup)
+
+		if result == const.ENoError:
+			self.__remove_local_on_success(localpath)
+		return result
 
 	def upload(self, localpath = '', remotepath = '', ondup = "overwrite"):
 		''' Usage: upload [localpath] [remotepath] [ondup] - \
@@ -1822,7 +1887,6 @@ try to create a file at PCS by combining slices, having MD5s specified
 
 	def __down_aria2c(self, remotefile, localfile):
 		url = "{}{}".format(dpcsurl, "file")
-
 		# i think encoding in UTF-8 before escaping is presumably the best practice
 		# http://stackoverflow.com/a/913653/404271
 		pars = {
@@ -1830,19 +1894,32 @@ try to create a file at PCS by combining slices, having MD5s specified
 			"path": remotefile.encode('utf-8'),
 			"access_token": self.__access_token,
 			}
-
 		full_url = "{}?{}".format(url, ulp.urlencode(pars))
-
 		cmd = ['aria2c', '--user-agent="{}"'.format(const.UserAgent)] \
 			+ shlex.split(self.__downloader_args) \
 			+ ['-o', localfile, full_url]
-		self.pd("call: {}".format(cmd))
-		ret = subprocess.call(cmd)
-		self.pd("aria2c exited with status: {}".format(ret))
-		# TODO: a finer map return codes to our internal errors
-		if ret != const.ENoError:
-			ret == const.ERequestFailed
-		return ret
+
+		tries = self.__retry
+		subret = 0
+		i = 0
+		while True:
+			self.pd("call: {}".format(cmd))
+			subret = subprocess.call(cmd)
+			self.pd("aria2c exited with status: {}".format(subret))
+			if subret == 0:
+				return const.ENoError
+			i += 1
+			if i < tries:
+				# algo changed: delay more after each failure
+				delay = const.RetryDelayInSec * i
+				perr("Waiting {} seconds before retrying...".format(delay))
+				time.sleep(delay)
+				perr("Request Try #{} / {}".format(i + 1, tries))
+			else:
+				perr("Maximum number ({}) of tries failed.".format(tries))
+				if self.__quit_when_fail:
+					self.quit(const.EMaxRetry)
+				return const.EMaxRetry
 
 	# requirment: self.__remote_json is already gotten
 	def __downchunks(self, rfile, start):
@@ -1904,12 +1981,12 @@ try to create a file at PCS by combining slices, having MD5s specified
 
 		return result
 
-
 	def __downfile(self, remotefile, localfile):
 		# TODO: this is a quick patch
 		if not self.__shallinclude(localfile, remotefile, False):
 			# since we are not going to download it, there is no error
-			return const.ENoError
+			#return const.ENoError
+			return const.ESkipped
 
 		result = const.ENoError
 		rfile = remotefile
@@ -1931,12 +2008,14 @@ try to create a file at PCS by combining slices, having MD5s specified
 			if const.ENoError == self.__verify_current_file(self.__remote_json, False) \
 				and not (self.__downloader[:5] == const.DownloaderAria2 and os.path.exists(localfile + '.aria2')):
 				self.pd("Same local file '{}' already exists, skip downloading".format(localfile))
+				self.__remove_remote_on_success(remotefile)
 				return const.ENoError
 			else:
 				if not self.shalloverwrite("Same-name locale file '{}' exists but is different, "
 						"do you want to overwrite it? [y/N]".format(localfile)):
 					pinfo("Same-name local file '{}' exists but is different, skip downloading".format(localfile))
-					return const.ENoError
+					#return const.ENoError
+					return const.ESkipped
 
 			if self.__resumedownload and \
 				self.__compare_size(self.__current_file_size, self.__remote_json) == 2:
@@ -1952,7 +2031,8 @@ try to create a file at PCS by combining slices, having MD5s specified
 			if not self.shalloverwrite("Same-name directory '{}' exists, "
 				"do you want to remove it? [y/N]".format(localfile)):
 				pinfo("Same-name directory '{}' exists, skip downloading".format(localfile))
-				return const.ENoError
+				#return const.ENoError
+				return const.ESkipped
 
 			self.pv("Directory with the same name '{}' exists, removing ...".format(localfile))
 			result = removedir(localfile, self.verbose)
@@ -1970,9 +2050,14 @@ try to create a file at PCS by combining slices, having MD5s specified
 				return result
 
 		if self.__downloader[:5] == const.DownloaderAria2:
-			return self.__down_aria2c(rfile, localfile)
+			result = self.__down_aria2c(rfile, localfile)
 		else:
-			return self.__downchunks(rfile, offset)
+			result = self.__downchunks(rfile, offset)
+
+		if result == const.ENoError:
+			self.__remove_remote_on_success(remotefile)
+
+		return result
 
 	def downfile(self, remotefile, localpath = ''):
 		''' Usage: downfile <remotefile> [localpath] - \
@@ -2144,10 +2229,14 @@ To stream a file, you can use the 'mkfifo' trick with omxplayer etc.:
 			"Directory Download")
 
 	def __downdir(self, rpath, lpath):
+		result = const.ENoError
 		if Pool and self.processes > 1:
-			return self.__downdir_multi(rpath, lpath)
+			result = self.__downdir_multi(rpath, lpath)
 		else:
-			return self.__downdir_single(rpath, lpath)
+			result = self.__downdir_single(rpath, lpath)
+		if result == const.ENoError:
+			self.__remove_remote_on_success(rpath)
+		return result
 
 	def downdir(self, remotepath = None, localpath = None):
 		''' Usage: downdir [remotedir] [localdir] - \
@@ -2199,7 +2288,8 @@ download a remote directory (recursively) / file
 		# the code still works because Baidu Yun doesn't require
 		# parent directory to exist remotely to upload / create a file
 		if not self.__shallinclude('.', rpath, True):
-			return const.ENoError
+			#return const.ENoError
+			return const.ESkipped
 
 		self.pd("Making remote directory '{}'".format(rpath))
 
@@ -2286,7 +2376,7 @@ copy a file / dir remotely at Baidu Yun
 		rid = j['request_id']
 		if rid:
 			pr("Deletion request '{}' OK".format(rid))
-			pr("Usage 'list' command to confirm")
+			pr("Use 'list' command to confirm")
 
 			return const.ENoError
 		else:
@@ -2670,6 +2760,8 @@ if not specified, it defaults to the root directory
 			if subresult != const.ENoError:
 				result = subresult
 
+		if result == const.ENoError:
+			self.__remove_remote_on_success(rpath)
 		return result
 
 	def __syncup_diff_one(self, rpath, localdir, d):
@@ -2791,6 +2883,8 @@ if not specified, it defaults to the root directory
 			subresult = self.__syncup_delete_remote(rpath, remote)
 			if subresult != const.ENoError:
 				result = subresult
+		if result == const.ENoError:
+			self.__remove_local_on_success(localdir)
 		return result
 
 	def dumpcache(self):
@@ -3381,8 +3475,11 @@ def getparser():
 		dest="resumedl_revertcount", default=const.DefaultResumeDlRevertCount,
 		type=int, metavar='RCOUNT',
 		help="Revert back at least %(metavar)s download chunk(s) and align to chunk boundary when resuming the download. A negative value means NO reverts. [default: %(default)s]")
+	parser.add_argument("--move",
+		dest="deletesource", action="store_true",
+		help="Delete source files/directories after download/upload/syncdown/syncup is successful (This will force verification of the files). [default: %(default)s]")
 	if Pool:
-		parser.add_argument("--processes",
+		parser.add_argument(const.MultiprocessOption,
 			dest="processes", default=const.DefaultProcessCount, type=int,
 			help="Number of parallel processes. (Only applies to dir sync/dl/ul). [default: %(default)s]")
 
@@ -3487,7 +3584,7 @@ def main(argv=None): # IGNORE:C0111
 		if os.path.exists(const.HashCachePath):
 			cachesize = getfilesize(const.HashCachePath)
 			if cachesize > 10 * const.OneM or cachesize == -1:
-				pr((
+				pwarn((
 				   "*** WARNING ***\n"
 				   "Hash Cache file '{0}' is very large ({1}).\n"
 				   "This may affect program's performance (high memory consumption).\n"
@@ -3504,13 +3601,13 @@ def main(argv=None): # IGNORE:C0111
 		try:
 			slice_size = interpret_size(args.slice)
 		except (ValueError, KeyError):
-			pr("Error: Invalid slice size specified '{}'".format(args.slice))
+			perr("Error: Invalid slice size specified '{}'".format(args.slice))
 			return const.EArgument
 
 		try:
 			chunk_size = interpret_size(args.chunk)
 		except (ValueError, KeyError):
-			pr("Error: Invalid slice size specified '{}'".format(args.slice))
+			perr("Error: Invalid slice size specified '{}'".format(args.slice))
 			return const.EArgument
 
 		if len(args.command) <= 0 or \
@@ -3539,6 +3636,7 @@ def main(argv=None): # IGNORE:C0111
 				'selectmirror': args.selectmirror,
 				'configdir': args.configdir,
 				'resumedl_revertcount': args.resumedl_revertcount,
+				'deletesource': args.deletesource,
 				'downloader': args.downloader,
 				'downloader_args': dl_args,
 				'verbose': args.verbose,
@@ -3557,8 +3655,11 @@ def main(argv=None): # IGNORE:C0111
 				else:
 					uargs.append(arg)
 			result = getattr(by, args.command[0])(*uargs)
+			if result != const.ENoError:
+				errmsg = '-' * 64 + "\nError {}{}".format(result, ': ' + const.ErrorExplanations[result] if result in const.ErrorExplanations else '')
+				perr(errmsg)
 		else:
-			pr("Error: Command '{}' not available.".format(args.command[0]))
+			perr("Error: Command '{}' not available.".format(args.command[0]))
 			parser.print_help()
 			return const.EParameter